IDMOB-10 Functionality for Apple account consolidation
When moving an app from one Apple account to another
it's not possible to keep the keychain data.
If the data is not there, a logged in user will not be logged in anymore.

To overcome this the move can be planned in 2 steps:
1. In the first step an app is upgraded on the old Apple account
  with a mode SPiDTokenStorageModeMigratePreITunesAccountMove to export keychain data to NSUserDefaults.
2. The next step is an app upgrade on to the new Apple account
  with a mode SPiDTokenStorageModeMigratePostITunesAccountMove to import data from NSUserDefaults.

Both versions should stay in the store for a while
until the most users are able to upgrade and run it.

Tests:
* SPiDTokenStorageTests
* SPiDTokenStorageUserDefaultsBackendTests